+++ benchmarks/duckdb-bench/src/lib.rs
@@ -16,7 +16,6 @@ use vortex_bench::Format;
 use vortex_bench::IdempotentPath;
 use vortex_bench::generate_duckdb_registration_sql;
 use vortex_bench::runner::BenchmarkQueryResult;
-use vortex_duckdb::duckdb::Config;
 use vortex_duckdb::duckdb::Connection;
 use vortex_duckdb::duckdb::Database;
 use vortex_duckdb::duckdb::QueryResult;
@@ -75,30 +74,21 @@ impl DuckClient {
         path: Option<PathBuf>,
         threads: Option<usize>,
     ) -> Result<(Database, Connection)> {
-        let mut config = Config::new().vortex_expect("failed to create duckdb config");
-
-        // Set DuckDB thread count if specified
-        if let Some(thread_count) = threads {
-            config.set("threads", &format!("{}", thread_count))?;
-        }
-
         let db = match path {
-            Some(path) => Database::open_with_config(path, config),
-            None => Database::open_in_memory_with_config(config),
+            Some(path) => Database::open(path),
+            None => Database::open_in_memory(),
         }?;
 
         let connection = db.connect()?;
         vortex_duckdb::initialize(&db)?;
 
-        // Enable Parquet metadata cache for all benchmark runs.
-        //
+        if let Some(thread_count) = threads {
+            connection.query(&format!("SET threads = {thread_count}"))?;
+        }
+
         // `parquet_metadata_cache` is an extension-specific option that's
         // only available after the Parquet extension is loaded. The Parquet
         // extension is loaded after the connection is established.
-        //
-        // Passing the option to `open_with_config` before leads to
-        // "Invalid Input Error: The following options were not recognized:
-        // parquet_metadata_cache" when running DuckDB in debug mode.
         connection.query("SET parquet_metadata_cache = true")?;
 
         Ok((db, connection))
