--- vortex-duckdb/cpp/include/duckdb_vx.h
@@ -1,23 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: Copyright the Vortex contributors
-
-#pragma once
-
-#include "duckdb_vx/client_context.h"
-#include "duckdb_vx/optimizer.h"
-#include "duckdb_vx/config.h"
-#include "duckdb_vx/copy_function.h"
-#include "duckdb_vx/data.h"
-#include "duckdb_vx/data_chunk.h"
-#include "duckdb_vx/error.h"
-#include "duckdb_vx/expr.h"
-#include "duckdb_vx/file_system.h"
-#include "duckdb_vx/logical_type.h"
-#include "duckdb_vx/reusable_dict.h"
-#include "duckdb_vx/replacement_scan.h"
-#include "duckdb_vx/scalar_function.h"
-#include "duckdb_vx/table_filter.h"
-#include "duckdb_vx/table_function.h"
-#include "duckdb_vx/value.h"
-#include "duckdb_vx/vector.h"
-#include "duckdb_vx/vector_buffer.h"
