+++ vortex-duckdb/cpp/include/vector.h
@@ -1,28 +1,14 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
-
 #pragma once
+#include "vortex_duckdb.h"
 
-#include "duckdb.h"
-
-#include "duckdb_vx/data.h"
-#include "duckdb_vx/error.h"
-#include "duckdb_vx/vector_buffer.h"
-
-#ifdef __cplusplus /* If compiled as C++, use C ABI */
+#ifdef __cplusplus
 extern "C" {
 #endif
 
 // Create a vector that slices another vector between a pair of offsets [offset, end)
 duckdb_vector duckdb_vx_vector_slice(duckdb_vector ffi_vector, idx_t offset, idx_t end);
-
-/// Slice the vector to a new dictionary vector, using the current vector's values and
-/// the provided selection vector.
-///
-/// A dictionary slice holds a strong reference to all memory it uses.
-void duckdb_vx_vector_slice_to_dictionary(duckdb_vector ffi_vector,
-                                          duckdb_selection_vector selection_vector,
-                                          idx_t selection_vector_length);
 
 /// Creates a dictionary vector for a given values vector and selection vector.
 ///
@@ -38,6 +24,27 @@ void duckdb_vx_vector_dictionary(duckdb_vector ffi_vector,
 
 void duckdb_vx_set_dictionary_vector_length(duckdb_vector dict, unsigned int len);
 
+// Reset vector's validity mask to nullptr, making all vector's elements valid.
+// vector must not be a DictionaryVector or a SequenceVector
+void duckdb_vx_vector_set_all_valid(duckdb_vector ffi_vector);
+
+// Set the data pointer for the vector. This is the start of the values array in the vector.
+void duckdb_vx_vector_set_data_ptr(duckdb_vector ffi_vector, void *ptr);
+
+// Converts a duckdb flat vector into a Sequence vector.
+void duckdb_vx_sequence_vector(duckdb_vector c_vector, int64_t start, int64_t step, idx_t capacity);
+
+void duckdb_vector_flatten(duckdb_vector vector, unsigned long len);
+
+duckdb_value duckdb_vx_vector_get_value(duckdb_vector ffi_vector, idx_t index);
+
+typedef struct duckdb_vx_vector_buffer_ *duckdb_vx_vector_buffer;
+
+// Create a external vector buffer from an existing data buffer
+duckdb_vx_vector_buffer duckdb_vx_vector_buffer_create(duckdb_vx_data buffer);
+
+void duckdb_vx_vector_buffer_destroy(duckdb_vx_vector_buffer *buffer);
+
 // Add the buffer to the string vector (basically, keep it alive as long as the vector).
 void duckdb_vx_string_vector_add_vector_data_buffer(duckdb_vector ffi_vector, duckdb_vx_vector_buffer buffer);
 
@@ -45,13 +52,6 @@ void duckdb_vx_string_vector_add_vector_data_buffer(duckdb_vector ffi_vector, du
 // pointer. You must ensure that the ptr is valid for the lifetime of the vector and the ptr addr + size is
 // valid.
 void duckdb_vx_vector_set_vector_data_buffer(duckdb_vector ffi_vector, duckdb_vx_vector_buffer buffer);
-
-// Reset vector's validity mask to nullptr, making all vector's elements valid.
-// vector must not be a DictionaryVector or a SequenceVector
-void duckdb_vx_vector_set_all_valid(duckdb_vector ffi_vector);
-
-// Set the data pointer for the vector. This is the start of the values array in the vector.
-void duckdb_vx_vector_set_data_ptr(duckdb_vector ffi_vector, void *ptr);
 
 // Set the validity pointer for the vector to external data, and store the buffer in auxiliary
 // to keep it alive. The validity pointer is derived from data_ptr at the given u64 offset.
@@ -62,15 +62,6 @@ void duckdb_vx_vector_set_validity_data(duckdb_vector ffi_vector,
                                         duckdb_vx_vector_buffer buffer,
                                         void *data_ptr);
 
-// Converts a duckdb flat vector into a Sequence vector.
-void duckdb_vx_sequence_vector(duckdb_vector c_vector, int64_t start, int64_t step, idx_t capacity);
-
-void duckdb_vector_flatten(duckdb_vector vector, unsigned long len);
-
-const char *duckdb_vector_to_string(duckdb_vector vector, unsigned long len, duckdb_vx_error *err);
-
-duckdb_value duckdb_vx_vector_get_value(duckdb_vector ffi_vector, idx_t index);
-
-#ifdef __cplusplus /* End C ABI */
+#ifdef __cplusplus
 }
 #endif
