+++ vortex-duckdb/cpp/vector.cpp
@@ -1,12 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-#include "include/duckdb_vx/vector.h"
-#include "duckdb/common/vector.hpp"
+#include "vector.h"
+#include "vector.hpp"
+
 #include "duckdb/common/types/value.hpp"
 #include "duckdb/common/types/vector.hpp"
-
-#include "duckdb_vx/vector_buffer.hpp"
 
 using namespace duckdb;
 
@@ -14,14 +13,6 @@ extern "C" duckdb_vector duckdb_vx_vector_slice(duckdb_vector ffi_vector, idx_t 
     const Vector &vector = *reinterpret_cast<Vector *>(ffi_vector);
     Vector *const sliced = new Vector(vector, offset, end);
     return reinterpret_cast<duckdb_vector>(sliced);
-}
-
-extern "C" void duckdb_vx_vector_slice_to_dictionary(duckdb_vector ffi_vector,
-                                                     duckdb_selection_vector ffi_sel_vec,
-                                                     idx_t selection_vector_length) {
-    auto vector = reinterpret_cast<Vector *>(ffi_vector);
-    auto sel_vec = reinterpret_cast<SelectionVector *>(ffi_sel_vec);
-    vector->Slice(*sel_vec, selection_vector_length);
 }
 
 extern "C" void duckdb_vx_vector_dictionary(duckdb_vector ffi_vector,
@@ -46,10 +37,10 @@ duckdb_vx_sequence_vector(duckdb_vector c_vector, int64_t start, int64_t step, i
     vector->Sequence(start, step, capacity);
 }
 
-namespace vortex {
-
 // This is a complete hack to access the data buffer and pointer of a vector.
-class DataVector : public Vector {
+// Duckdb passes us Vectors and not VortexVectors. This only works because
+// VortexVector doesn't add any members.
+class VortexVector final : public Vector {
 public:
     inline void SetDataBuffer(buffer_ptr<VectorBuffer> new_buffer) {
         buffer = std::move(new_buffer);
@@ -63,9 +54,10 @@ public:
         return validity;
     };
 };
+static_assert(sizeof(VortexVector) == sizeof(Vector));
 
 // Same hack for ValidityMask: access protected fields via inheritance.
-class ExternalValidityMask : public ValidityMask {
+class ExternalValidityMask final : public ValidityMask {
 public:
     inline void SetExternal(idx_t u64_offset, idx_t cap, buffer_ptr<ValidityBuffer> keeper) {
         validity_data = std::move(keeper);
@@ -74,27 +66,26 @@ public:
         capacity = cap;
     };
 };
-
-} // namespace vortex
+static_assert(sizeof(ExternalValidityMask) == sizeof(ValidityMask));
 
 extern "C" void duckdb_vx_string_vector_add_vector_data_buffer(duckdb_vector ffi_vector,
                                                                duckdb_vx_vector_buffer buffer) {
     auto vector = reinterpret_cast<Vector *>(ffi_vector);
-    auto data = reinterpret_cast<shared_ptr<vortex::ExternalVectorBuffer> *>(buffer);
+    auto data = reinterpret_cast<shared_ptr<ExternalVectorBuffer> *>(buffer);
     StringVector::AddBuffer(*vector, *data);
 }
 
 extern "C" void duckdb_vx_vector_set_vector_data_buffer(duckdb_vector ffi_vector,
                                                         duckdb_vx_vector_buffer buffer) {
     auto vector = reinterpret_cast<Vector *>(ffi_vector);
-    auto dvector = reinterpret_cast<vortex::DataVector *>(vector);
-    auto data = reinterpret_cast<shared_ptr<vortex::ExternalVectorBuffer> *>(buffer);
+    auto dvector = reinterpret_cast<VortexVector *>(vector);
+    auto data = reinterpret_cast<shared_ptr<ExternalVectorBuffer> *>(buffer);
     dvector->SetDataBuffer(*data);
 }
 
 extern "C" void duckdb_vx_vector_set_data_ptr(duckdb_vector ffi_vector, void *ptr) {
     auto vector = reinterpret_cast<Vector *>(ffi_vector);
-    auto dvector = reinterpret_cast<vortex::DataVector *>(vector);
+    auto dvector = reinterpret_cast<VortexVector *>(vector);
     dvector->SetDataPtr((data_ptr_t)ptr);
 }
 
@@ -103,16 +94,16 @@ extern "C" void duckdb_vx_vector_set_validity_data(duckdb_vector ffi_vector,
                                                    idx_t capacity,
                                                    duckdb_vx_vector_buffer buffer,
                                                    void *data_ptr) {
-    auto dvector = reinterpret_cast<vortex::DataVector *>(ffi_vector);
+    auto dvector = reinterpret_cast<VortexVector *>(ffi_vector);
     auto &validity = dvector->GetValidity();
     // ExternalValidityMask adds no members, so this downcast only exposes
     // access to ValidityMask's protected fields.
-    auto ext_validity = static_cast<vortex::ExternalValidityMask *>(&validity);
+    auto ext_validity = static_cast<ExternalValidityMask *>(&validity);
 
     // Use the shared_ptr aliasing constructor: the control block ref-counts the
     // ExternalVectorBuffer (preventing the Rust buffer from being freed),
     // while the stored pointer points to the explicit data_ptr.
-    auto ext_buf = reinterpret_cast<shared_ptr<vortex::ExternalVectorBuffer> *>(buffer);
+    auto ext_buf = reinterpret_cast<shared_ptr<ExternalVectorBuffer> *>(buffer);
     auto keeper = shared_ptr<TemplatedValidityData<validity_t>>(
         *ext_buf,
         reinterpret_cast<TemplatedValidityData<validity_t> *>(data_ptr));
@@ -132,21 +123,6 @@ void duckdb_vector_flatten(duckdb_vector vector, unsigned long len) {
     dvector->Flatten(len);
 }
 
-const char *duckdb_vector_to_string(duckdb_vector vector, unsigned long len, duckdb_vx_error *err) {
-    try {
-        auto dvector = reinterpret_cast<Vector *>(vector);
-        auto str = dvector->ToString(len);
-        auto result = static_cast<char *>(duckdb_malloc(str.size() + 1));
-        memcpy(result, str.c_str(), str.size() + 1);
-        *err = nullptr;
-        return result;
-    } catch (std::runtime_error &e) {
-        auto s = e.what();
-        *err = duckdb_vx_error_create(s, strlen(s));
-        return nullptr;
-    }
-}
-
 void duckdb_vx_vector_set_all_valid(duckdb_vector ffi_vector) {
     using enum VectorType;
     Vector &vector = *reinterpret_cast<Vector *>(ffi_vector);
@@ -161,5 +137,20 @@ void duckdb_vx_vector_set_all_valid(duckdb_vector ffi_vector) {
         return FSSTVector::Validity(vector).Reset();
     default:
         __builtin_unreachable();
+    }
+}
+
+extern "C" duckdb_vx_vector_buffer duckdb_vx_vector_buffer_create(duckdb_vx_data buffer) {
+    auto data = reinterpret_cast<CData *>(buffer);
+    auto *shared_buffer =
+        new shared_ptr<ExternalVectorBuffer>(make_shared_ptr<ExternalVectorBuffer>(unique_ptr<CData>(data)));
+    return reinterpret_cast<duckdb_vx_vector_buffer>(shared_buffer);
+}
+
+extern "C" void duckdb_vx_vector_buffer_destroy(duckdb_vx_vector_buffer *buffer) {
+    if (buffer != nullptr && *buffer != nullptr) {
+        auto shared_buffer = reinterpret_cast<shared_ptr<ExternalVectorBuffer> *>(*buffer);
+        delete shared_buffer;
+        *buffer = nullptr;
     }
 }
