+++ vortex-duckdb/src/duckdb/ddb_string.rs
@@ -6,8 +6,6 @@ use std::fmt::Display;
 use std::fmt::Formatter;
 
 use vortex::dtype::FieldName;
-use vortex::error::VortexExpect;
-use vortex::error::vortex_err;
 
 use crate::cpp::duckdb_free;
 use crate::lifetime_wrapper;
@@ -18,21 +16,6 @@ lifetime_wrapper!(
     *mut std::ffi::c_char,
     |ptr: &mut *mut std::ffi::c_char| unsafe { duckdb_free((*ptr).cast()) }
 );
-
-impl DDBString {
-    /// Creates an owned DDBString from a C string pointer, validating it is UTF-8.
-    ///
-    /// # Safety
-    ///
-    /// The pointer must be a valid, non-null, null-terminated C string allocated by DuckDB.
-    pub unsafe fn from_c_str(ptr: *mut std::ffi::c_char) -> Self {
-        unsafe { CStr::from_ptr(ptr) }
-            .to_str()
-            .map_err(|e| vortex_err!("Failed to convert C string to str: {e}"))
-            .vortex_expect("DuckDB string should be valid UTF-8");
-        unsafe { Self::own(ptr) }
-    }
-}
 
 impl Display for DDBString {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
