+++ vortex-duckdb/src/duckdb/scalar_function.rs
@@ -5,8 +5,6 @@ use vortex::error::VortexExpect;
 use vortex::error::vortex_err;
 
 use crate::cpp;
-use crate::duckdb::LogicalType;
-use crate::duckdb::LogicalTypeRef;
 use crate::lifetime_wrapper;
 
 lifetime_wrapper!(ScalarFunction, cpp::duckdb_vx_sfunc, |_| {});
@@ -20,9 +18,5 @@ impl ScalarFunctionRef {
                 .map_err(|e| vortex_err!("invalid utf-8: {e}"))
                 .vortex_expect("scalar function name should be valid UTF-8")
         }
-    }
-
-    pub fn return_type(&self) -> &LogicalTypeRef {
-        unsafe { LogicalType::borrow(cpp::duckdb_vx_sfunc_return_type(self.as_ptr())) }
     }
 }
